work on the original poly in layer instead to make sure line changes are also properly applied
